Base commit
This commit provides the base code, dataset, and readme. Dataset is from Lloyd et al. (in prep) can be used as the example data. The code requires the presence of the following headers  in the CSV file but does not need to be in any specific order: U238, Yb172, Th2322, Lu175, Nd146, Ce140, La139, Pr141, Sm147, Eu153, Sm147, Gd157, Y89, FilteredAge, lambda1, lambda2 and lambda3